requirements.txt
# Django Core
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0

# Database & Environment
psycopg2-binary>=2.9.6
django-environ>=0.10.0

# API Enhancement
django-filter>=23.2
drf-spectacular>=0.26.2

# Utilities
django-cors-headers>=4.0.0
Pillow>=9.5.0
python-dateutil>=2.8.2

# Development/Testing
pytest-django>=4.5.2
factory-boy>=3.2.1
